Adjusted ownership of connector, now it follows how the architecture diagram displayed it
