+++ Dashboard/Models/DataStore.cs
@@ -1,14 +1,25 @@
 ﻿using System;
+using Dashboard.Connectors;
 
 namespace Dashboard.Models;
 
-public class DataStore : IDataStore
+public class DataStore : IDataStore, IDisposable
 {
     public event EventHandler? DataUpdated;
 
     public double Speed { get; private set; }
     public double SteeringAngle { get; private set; }
     public double BrakePressure { get; private set; }
+
+    private readonly IConnector _connector;
+
+    public DataStore(IConnector connector)
+    {
+        _connector = connector;
+        _connector.DataUpdated += OnDataUpdated;
+
+        _connector.Start();
+    }
 
     public void UpdateData(double speed, double steeringAngle, double brakePressure)
     {
@@ -17,5 +28,19 @@ public class DataStore : IDataStore
         BrakePressure = brakePressure;
 
         DataUpdated?.Invoke(this, EventArgs.Empty);
+    }
+
+    private void OnDataUpdated(object? sender, DataUpdatedEventArgs e)
+    {
+        UpdateData(e.Speed, e.SteeringAngle, e.BrakePressure);
+    }
+
+    public void Dispose()
+    {
+        _connector.DataUpdated -= OnDataUpdated;
+
+        _connector.Stop();
+
+        GC.SuppressFinalize(this);
     }
 }
