+++ Dashboard/ViewModels/MainViewModel.cs
@@ -4,41 +4,25 @@ using Dashboard.Models;
 
 namespace Dashboard.ViewModels;
 
-public class MainViewModel : ViewModelBase
+public class MainViewModel : ViewModelBase, IDisposable
 {
     private readonly IDataStore _dataStore;
-    private readonly IConnector _connector;
 
     public double Speed => _dataStore.Speed;
     public double SteeringAngle => _dataStore.SteeringAngle;
     public double BrakePressure => _dataStore.BrakePressure;
 
-    public MainViewModel(IDataStore dataStore, IConnector connector)
+    public MainViewModel(IDataStore dataStore)
     {
         _dataStore = dataStore;
-        _connector = connector;
 
         _dataStore.DataUpdated += OnDataChanged;
-        _connector.DataUpdated += OnDataUpdated;
-
-        _connector.Start();
     }
 
     public MainViewModel()
     {
         // This constructor is used for design-time data, so we don't need to start the connector
-        _dataStore = new DataStore();
-        _connector = new DummyConnector();
-    }
-
-    /// <summary>
-    ///     Notifies the data store that new data has arrived.
-    /// </summary>
-    /// <param name="sender">The sender of the event</param>
-    /// <param name="e">The event containing the new data</param>
-    private void OnDataUpdated(object? sender, DataUpdatedEventArgs e)
-    {
-        _dataStore.UpdateData(e.Speed, e.SteeringAngle, e.BrakePressure);
+        _dataStore = new DataStore(new DummyConnector());
     }
 
     /// <summary>
@@ -57,8 +41,9 @@ public class MainViewModel : ViewModelBase
     public void Dispose()
     {
         _dataStore.DataUpdated -= OnDataChanged;
-        _connector.DataUpdated -= OnDataUpdated;
 
-        _connector.Stop();
+        _dataStore.Dispose();
+
+        GC.SuppressFinalize(this);
     }
 }
