+++ Dashboard/Views/MainWindow.axaml.cs
@@ -15,6 +15,8 @@ public partial class MainWindow : Window
     {
         base.OnClosed(e);
 
+        // Dispose the view model when the window is closed, to clean up the data store and connector
+        // Otherwise, the connector thread will keep running in the background
         if (DataContext is MainViewModel model) model.Dispose();
     }
 }
